Trait console_draw::ConsoleCanvas

pub trait ConsoleCanvas {
    fn draw_char(&mut self, x: uint, y: uint, c: char);
    fn clear(&mut self);
    fn add_modifier(&mut self, modifier: &Modifier);
    fn clear_modifiers(&mut self);
    fn supports_custom_colors(&self) -> bool;
    fn present(&mut self);
    fn cursor(&mut self, x: uint, y: uint);
    fn width(&self) -> uint;
    fn height(&self) -> uint;

    fn draw<S: Str>(&mut self, x: uint, y: uint, text: S) { ... }
    fn with_modifiers(&mut self, modifiers: &[Modifier], f: |&mut Self|) { ... }
}

Required Methods

fn draw_char(&mut self, x: uint, y: uint, c: char)

Draws a character to the console at the specified x, y position.

fn clear(&mut self)

Clears the screen without any garantee as to what color the screen is cleared with.

fn add_modifier(&mut self, modifier: &Modifier)

Adds a modifier to the console for all text printed afterwards.

fn clear_modifiers(&mut self)

Removes all modifiers from the console, returning it back to the default state.

fn supports_custom_colors(&self) -> bool

Returns true if this console has support for custom colors instead of the basic black, red, green, etc. If this method returns false, calling add_modifier with a custom color has unspecified behavior.

fn present(&mut self)

Swaps the back and front buffer and displays the text that has been printed to the screen using draw_char or draw.

fn cursor(&mut self, x: uint, y: uint)

Places the terminal cursor at the x, y position.

fn width(&self) -> uint

Returns the width of the console.

fn height(&self) -> uint

Returns the height of the console.

Provided Methods

fn draw<S: Str>(&mut self, x: uint, y: uint, text: S)

Draws a string to the console starting at position x, y and continuing to the right.

fn with_modifiers(&mut self, modifiers: &[Modifier], f: |&mut Self|)

With a set of modifiers, returns a closure that has a new Console that is has the modifiers set for the duration of the closure. Afterwards the modifiers state is cleared.